requirements.txt
torch>=2.1.2
torchvision>=0.16.2
torchaudio>=2.1.2
numpy>=1.26.3
python-chess>=1.999
